requirements.txt
django-treebeard==4.1.0
FeinCMS==1.13.1
